requirements.txt
requests
openai
dotenv
time 
csv
snowflake-connector-python

